Updated Home.htmls navbar
I have updated the navbar per the clients request now users can see that instead of a black navbar it is a white one, furthermore I have inputed a form for users wanting to purchase items in the store when users select a product the cart will update the number so if a user selects 3 items the cart will show the number 3. these changes will be included for the navbar on all of the HTML files.